Update README.md
Hi! Small suggestion to make it easier for other newbies. I was confused for a while that nothing was printed when the input wasn't toggling. 

Also, just so you know, we're hoping to use this code in a satellite to detect the frequency of a ring oscillator we made in open source silicon!